+++ src/Reface.App.Starter/AppContainers/AutofacContainerComponentContainer.cs
@@ -93,6 +93,7 @@ namespace Reface.AppStarter.AppContainers
             using (var scope = this.Container.BeginLifetimeScope("OnAppStarted"))
             {
                 IEventBus eventBus = scope.Resolve<IEventBus>();
+                eventBus.Publish(new AppStartingEvent(this, app));
                 eventBus.Publish(new AppStartedEvent(this, app));
             }
         }