+++ src/Reface.AppStarter.Tests/Repositories/IRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace Reface.AppStarter.Tests.Repositories
+{
+    public interface IRepository<TEntity>
+    {
+    }
+}