+++ src/Reface.AppStarter.Tests/Repositories/Repository.cs
@@ -1,0 +1,9 @@
+﻿using Reface.AppStarter.Attributes;
+
+namespace Reface.AppStarter.Tests.Repositories
+{
+    [Component]
+    public class Repository<TEntity>:IRepository<TEntity>
+    {
+    }
+}